feat: table works waw
