+++ src/pages/Main/BetterTable/index.jsx
@@ -1,297 +1,21 @@
 import React from 'react';
 import './styles.css';
-import PropTypes from 'prop-types';
-import clsx from 'clsx';
-import Status from 'components/status';
-import { lighten, makeStyles } from '@material-ui/core/styles';
-// import MaterialTable from 'material-table'
-// import Cell from './Cell';
-import hebrewLocalization from 'config/tableHebrew';
-import DeleteDialog from 'components/Confirm';
-import tableIcons from 'config/tableIcons';
-import DeleteIcon from '@material-ui/icons/Delete';
-import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
-
+import { makeStyles } from '@material-ui/core/styles';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
 import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import Paper from '@material-ui/core/Paper';
 
-import Toolbar from '@material-ui/core/Toolbar';
-import TableSortLabel from '@material-ui/core/TableSortLabel';
-import Box from '@material-ui/core/Box';
-import Collapse from '@material-ui/core/Collapse';
-import IconButton from '@material-ui/core/IconButton';
-import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
-import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
-import Typography from '@material-ui/core/Typography';
-import Checkbox from '@material-ui/core/Checkbox';
-import Tooltip from '@material-ui/core/Tooltip';
-import FormControlLabel from '@material-ui/core/FormControlLabel';
-import Switch from '@material-ui/core/Switch';
-import FilterListIcon from '@material-ui/icons/FilterList';
+import CustomTableRow from './Row';
+import CustomToolBar from './ToolBar';
+import CustomTableHead from './Head';
+import { getComparator, stableSort } from 'utils/tableUtils';
+
 import TablePagination from '@material-ui/core/TablePagination';
 
-import Cell from '../Table/Cell';
-
-
-const mock = {
-    id: 125435435345,
-    status: {
-        progress: 'completed',
-        step: 'target',
-        subStep: 'הסתיים',
-    },
-    primaryDomainUser: 'ads',
-    hierarchy: 'חברת הירקות/טיפשים/צוות חציל',
-    shadowUsers: ['es', 'target'],
-    fullName: 'חציל אפוי',
-    identifier: 1234521,
-    startDate: '12/02/1421',
-    endDate: '02/08/0002'
-};
-
-function Row(props) {
-    const { row, isItemSelected, labelId, handleClick } = props;
-    const [open, setOpen] = React.useState(false);
-    // const classes = useRowStyles();
-
-    //   <TableRow
-    //     hover
-    //     onClick={(event) => handleClick(event, row.name)}
-    //     role="checkbox"
-    //     aria-checked={isItemSelected}
-    //     tabIndex={-1}
-    //     key={row.name}
-    //     selected={isItemSelected}
-    //   >
-    //     <TableCell padding="checkbox">
-    //       <Checkbox
-    //         checked={isItemSelected}
-    //         inputProps={{ 'aria-labelledby': labelId }}
-    //       />
-    //     </TableCell>
-    //     <TableCell component="th" id={labelId} scope="row" padding="none">
-    //       {row.name}
-    //     </TableCell>
-    //     <TableCell align="right">{row.calories}</TableCell>
-    //     <TableCell align="right">{row.fat}</TableCell>
-    //     <TableCell align="right">{row.carbs}</TableCell>
-    //     <TableCell align="right">{row.protein}</TableCell>
-    //   </TableRow>
-
-    return (
-        <React.Fragment>
-            <TableRow className='' hover
-                onClick={(event) => handleClick(event, row.name)}
-                role="checkbox"
-                aria-checked={isItemSelected}
-                tabIndex={-1}
-                key={row.name}
-                selected={isItemSelected}>
-                <TableCell padding="checkbox">
-                    <Checkbox
-                        checked={isItemSelected}
-                        inputProps={{ 'aria-labelledby': labelId }}
-                    />
-                </TableCell>
-                <TableCell>
-                    <IconButton aria-label="expand row" size="small" onClick={(e) => { e.stopPropagation(); setOpen(!open) }}>
-                        {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-                    </IconButton>
-                </TableCell>
-                <TableCell component="th" scope="row">
-                    {row.name}
-                </TableCell>
-                <TableCell align="right">{row.calories}</TableCell>
-                <TableCell align="right">{row.fat}</TableCell>
-                <TableCell align="right">{row.carbs}</TableCell>
-                <TableCell align="right">{row.protein}</TableCell>
-            </TableRow>
-            <TableRow>
-                <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-                    <Collapse in={open} timeout="auto" unmountOnExit>
-                        <Box margin={1}>
-                            <Cell rowData={mock} />
-                        </Box>
-                    </Collapse>
-                </TableCell>
-            </TableRow>
-        </React.Fragment>
-    );
-}
-
-
-function createData(name, calories, fat, carbs, protein) {
-    return { name, calories, fat, carbs, protein };
-}
-
-const rows = [
-    createData('Cupcake', 305, 3.7, 67, 4.3),
-    createData('Donut', 452, 25.0, 51, 4.9),
-    createData('Eclair', 262, 16.0, 24, 6.0),
-    createData('Frozen yoghurt', 159, 6.0, 24, 4.0),
-    createData('Gingerbread', 356, 16.0, 49, 3.9),
-    createData('Honeycomb', 408, 3.2, 87, 6.5),
-    createData('Ice cream sandwich', 237, 9.0, 37, 4.3),
-    createData('Jelly Bean', 375, 0.0, 94, 0.0),
-    createData('KitKat', 518, 26.0, 65, 7.0),
-    createData('Lollipop', 392, 0.2, 98, 0.0),
-    createData('Marshmallow', 318, 0, 81, 2.0),
-    createData('Nougat', 360, 19.0, 9, 37.0),
-    createData('Oreo', 437, 18.0, 63, 4.0),
-];
-
-function descendingComparator(a, b, orderBy) {
-    if (b[orderBy] < a[orderBy]) {
-        return -1;
-    }
-    if (b[orderBy] > a[orderBy]) {
-        return 1;
-    }
-    return 0;
-}
-
-function getComparator(order, orderBy) {
-    return order === 'desc'
-        ? (a, b) => descendingComparator(a, b, orderBy)
-        : (a, b) => -descendingComparator(a, b, orderBy);
-}
-
-function stableSort(array, comparator) {
-    const stabilizedThis = array.map((el, index) => [el, index]);
-    stabilizedThis.sort((a, b) => {
-        const order = comparator(a[0], b[0]);
-        if (order !== 0) return order;
-        return a[1] - b[1];
-    });
-    return stabilizedThis.map((el) => el[0]);
-}
-
-const headCells = [
-    { id: 'name', numeric: false, disablePadding: true, label: 'Dessert (100g serving)' },
-    { id: 'calories', numeric: true, disablePadding: false, label: 'Calories' },
-    { id: 'fat', numeric: true, disablePadding: false, label: 'Fat (g)' },
-    { id: 'carbs', numeric: true, disablePadding: false, label: 'Carbs (g)' },
-    { id: 'protein', numeric: true, disablePadding: false, label: 'Protein (g)' },
-];
-
-function EnhancedTableHead(props) {
-    const { classes, onSelectAllClick, order, orderBy, numSelected, rowCount, onRequestSort } = props;
-    const createSortHandler = (property) => (event) => {
-        onRequestSort(event, property);
-    };
-
-    return (
-        <TableHead>
-            <TableRow>
-                <TableCell padding="checkbox">
-                    <Checkbox
-                        indeterminate={numSelected > 0 && numSelected < rowCount}
-                        checked={rowCount > 0 && numSelected === rowCount}
-                        onChange={onSelectAllClick}
-                        inputProps={{ 'aria-label': 'select all desserts' }}
-                    />
-                </TableCell>
-                {headCells.map((headCell) => (
-                    <TableCell
-                        key={headCell.id}
-                        align={headCell.numeric ? 'right' : 'left'}
-                        padding={headCell.disablePadding ? 'none' : 'default'}
-                        sortDirection={orderBy === headCell.id ? order : false}
-                    >
-                        <TableSortLabel
-                            active={orderBy === headCell.id}
-                            direction={orderBy === headCell.id ? order : 'asc'}
-                            onClick={createSortHandler(headCell.id)}
-                        >
-                            {headCell.label}
-                            {orderBy === headCell.id ? (
-                                <span className={classes.visuallyHidden}>
-                                    {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                                </span>
-                            ) : null}
-                        </TableSortLabel>
-                    </TableCell>
-                ))}
-            </TableRow>
-        </TableHead>
-    );
-}
-
-EnhancedTableHead.propTypes = {
-    classes: PropTypes.object.isRequired,
-    numSelected: PropTypes.number.isRequired,
-    onRequestSort: PropTypes.func.isRequired,
-    onSelectAllClick: PropTypes.func.isRequired,
-    order: PropTypes.oneOf(['asc', 'desc']).isRequired,
-    orderBy: PropTypes.string.isRequired,
-    rowCount: PropTypes.number.isRequired,
-};
-
-const useToolbarStyles = makeStyles((theme) => ({
-    root: {
-        paddingLeft: theme.spacing(2),
-        paddingRight: theme.spacing(1),
-    },
-    highlight:
-        theme.palette.type === 'light'
-            ? {
-                color: theme.palette.secondary.main,
-                backgroundColor: lighten(theme.palette.secondary.light, 0.85),
-            }
-            : {
-                color: theme.palette.text.primary,
-                backgroundColor: theme.palette.secondary.dark,
-            },
-    title: {
-        flex: '1 1 100%',
-    },
-}));
-
-const EnhancedTableToolbar = (props) => {
-    const classes = useToolbarStyles();
-    const { numSelected } = props;
-
-    return (
-        <Toolbar
-            className={clsx(classes.root, {
-                [classes.highlight]: numSelected > 0,
-            })}
-        >
-            {numSelected > 0 ? (
-                <Typography className={classes.title} color="inherit" variant="subtitle1" component="div">
-                    {numSelected} selected
-                </Typography>
-            ) : (
-                    <Typography className={classes.title} variant="h6" id="tableTitle" component="div">
-                        Nutrition
-                    </Typography>
-                )}
-
-            {numSelected > 0 ? (
-                <Tooltip title="Delete">
-                    <IconButton aria-label="delete">
-                        <DeleteIcon />
-                    </IconButton>
-                </Tooltip>
-            ) : (
-                    <Tooltip title="Filter list">
-                        <IconButton aria-label="filter list">
-                            <FilterListIcon />
-                        </IconButton>
-                    </Tooltip>
-                )}
-        </Toolbar>
-    );
-};
-
-EnhancedTableToolbar.propTypes = {
-    numSelected: PropTypes.number.isRequired,
-};
+import TextField from '@material-ui/core/TextField';
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -317,14 +41,18 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-export default function EnhancedTable() {
+export default function EnhancedTable({data = []}) {
     const classes = useStyles();
     const [order, setOrder] = React.useState('asc');
     const [orderBy, setOrderBy] = React.useState('calories');
     const [selected, setSelected] = React.useState([]);
     const [page, setPage] = React.useState(0);
-    const [dense, setDense] = React.useState(false);
     const [rowsPerPage, setRowsPerPage] = React.useState(5);
+    const [rows, setRows] = React.useState(data);
+    
+    React.useEffect(()=> {
+        setRows(data);
+    },[data]);
 
     const handleRequestSort = (event, property) => {
         const isAsc = orderBy === property && order === 'asc';
@@ -334,7 +62,7 @@ export default function EnhancedTable() {
 
     const handleSelectAllClick = (event) => {
         if (event.target.checked) {
-            const newSelecteds = rows.map((n) => n.name);
+            const newSelecteds = rows.map((n) => n.id);
             setSelected(newSelecteds);
             return;
         }
@@ -370,26 +98,57 @@ export default function EnhancedTable() {
         setPage(0);
     };
 
-    const handleChangeDense = (event) => {
-        setDense(event.target.checked);
-    };
+    const isSelected = (id) => selected.indexOf(id) !== -1;
+    
+    const searchHandler = (e) => {
+        let msg = e.target.value;
+        // const newSelecteds = rows.map((n) => n.id);
 
-    const isSelected = (name) => selected.indexOf(name) !== -1;
-
+        // setSelected([rows[0].id])
+        let matchedRows = [];
+        if(msg) {
+            // for(let dataRow of data) {
+            //     let values = Object.values(dataRow);
+            //     for(let valueStr of values) {
+            //         if(String(valueStr).includes(msg)) {
+            //             matchedRows.push(dataRow);
+            //             break;
+            //         }
+            //     }
+            // }
+            matchedRows = data.filter((dataRow) => {
+                let values = Object.values(dataRow);
+                let found = false;
+                for(let value of values) {
+                    let term = String(value);
+                    term = term.toLowerCase();
+                    let filterValue = msg.toLowerCase();
+                    if(term.includes(filterValue)) found = true;
+                }
+                return found;
+            })
+        }
+        else {
+            matchedRows = data;
+        }
+        // console.log(matchedRows);
+        setRows(matchedRows);
+    }
     const emptyRows = rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage);
 
     return (
         <div className={classes.root}>
             <Paper className={classes.paper}>
-                <EnhancedTableToolbar numSelected={selected.length} />
-                <TableContainer>
+                <CustomToolBar numSelected={selected.length} />
+                <TableContainer  style={{maxHeight: 600}}>
                     <Table
                         className={classes.table}
                         aria-labelledby="tableTitle"
-                        size={dense ? 'small' : 'medium'}
+                        // size={dense ? 'small' : 'medium'}
                         aria-label="enhanced table"
+                        stickyHeader
                     >
-                        <EnhancedTableHead
+                        <CustomTableHead
                             classes={classes}
                             numSelected={selected.length}
                             order={order}
@@ -402,40 +161,15 @@ export default function EnhancedTable() {
                             {stableSort(rows, getComparator(order, orderBy))
                                 .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
                                 .map((row, index) => {
-                                    const isItemSelected = isSelected(row.name);
+                                    const isItemSelected = isSelected(row.id);
                                     const labelId = `enhanced-table-checkbox-${index}`;
 
                                     return (
-                                        //   <TableRow
-                                        //     hover
-                                        //     onClick={(event) => handleClick(event, row.name)}
-                                        //     role="checkbox"
-                                        //     aria-checked={isItemSelected}
-                                        //     tabIndex={-1}
-                                        //     key={row.name}
-                                        //     selected={isItemSelected}
-                                        //   >
-                                        //     <TableCell padding="checkbox">
-                                        //       <Checkbox
-                                        //         checked={isItemSelected}
-                                        //         inputProps={{ 'aria-labelledby': labelId }}
-                                        //       />
-                                        //     </TableCell>
-                                        //     <TableCell component="th" id={labelId} scope="row" padding="none">
-                                        //       {row.name}
-                                        //     </TableCell>
-                                        //     <TableCell align="right">{row.calories}</TableCell>
-                                        //     <TableCell align="right">{row.fat}</TableCell>
-                                        //     <TableCell align="right">{row.carbs}</TableCell>
-                                        //     <TableCell align="right">{row.protein}</TableCell>
-                                        //   </TableRow>
-                                        // rows.map((row) => (
-                                        <Row key={row.name} row={row} isItemSelected={isItemSelected} labelId={labelId} handleClick={handleClick} />
-                                        //   ))
+                                        <CustomTableRow key={row.id} row={row} isItemSelected={isItemSelected} labelId={labelId} handleClick={handleClick} />
                                     );
                                 })}
                             {emptyRows > 0 && (
-                                <TableRow style={{ height: (dense ? 33 : 53) * emptyRows }}>
+                                <TableRow style={{ height: (53) * emptyRows }}>
                                     <TableCell colSpan={6} />
                                 </TableRow>
                             )}
@@ -447,15 +181,18 @@ export default function EnhancedTable() {
                     component="div"
                     count={rows.length}
                     rowsPerPage={rowsPerPage}
+                    labelRowsPerPage='שורות בעמוד:'
+                    nextIconButtonText='עמוד הבא'
+                    backIconButtonText='עמוד הקודם'
+                    labelDisplayedRows={({from, to, count}) => 
+                        `${from}-${to} מתוך ${count}`
+                    }
                     page={page}
                     onChangePage={handleChangePage}
                     onChangeRowsPerPage={handleChangeRowsPerPage}
                 />
             </Paper>
-            <FormControlLabel
-                control={<Switch checked={dense} onChange={handleChangeDense} />}
-                label="Dense padding"
-            />
+            <TextField id="outlined-basic" label="Outlined" variant="outlined" onChange={searchHandler} />
         </div>
     );
 }
